Split out request handlers
• Move handlers to dedicated files
• Improve code structure and readability
• Create new handlers directory
• Extract handler classes from server.ts
• Update version to 0.2.0
